texture_cache: image upload logic moved into `Image` object
